+++ tauri-app/src/lib/components/ModalVaultDeposit.svelte
@@ -3,7 +3,7 @@
   import type { TokenVault as TokenVaultDetail } from '$lib/typeshare/vaultDetail';
   import type { TokenVault as TokenVaultListItem } from '$lib/typeshare/vaultsList';
   import InputTokenAmount from '$lib/components/InputTokenAmount.svelte';
-  import { vaultDeposit } from '$lib/utils/vaultDeposit';
+  import { vaultDeposit } from '$lib/services/vaultDeposit';
   import { bigintStringToHex } from '$lib/utils/hex';
   import ButtonLoading from '$lib/components/ButtonLoading.svelte';
