+++ tauri-app/src/lib/types/hash.ts
