+++ tauri-app/src/routes/orders/[id]/+page.svelte
@@ -7,11 +7,11 @@
   import BadgeActive from '$lib/components/BadgeActive.svelte';
   import { formatTimestampSecondsAsLocal, timestampSecondsToUTCTimestamp } from '$lib/utils/time';
   import ButtonVaultLink from '$lib/components/ButtonVaultLink.svelte';
-  import { orderRemove } from '$lib/utils/orderRemove';
+  import { orderRemove } from '$lib/services/orderRemove';
   import PageHeader from '$lib/components/PageHeader.svelte';
   import { page } from '$app/stores';
   import Hash from '$lib/components/Hash.svelte';
-  import { HashType } from '$lib/utils/hash';
+  import { HashType } from '$lib/types/hash';
   import AppTable from '$lib/components/AppTable.svelte';
   import { sortBy } from 'lodash';
   import { useOrderTakesList } from '$lib/stores/orderTakesList';
